feat : CustomerRepository 인터페이스 도입 및 분리
