+++ src/main/java/com/zerozae/voucher/domain/customer/Customer.java
@@ -1,5 +1,6 @@
 package com.zerozae.voucher.domain.customer;
 
+import com.zerozae.voucher.dto.customer.CustomerRequest;
 import com.zerozae.voucher.exception.ErrorMessage;
 import lombok.Getter;
 
@@ -10,14 +11,20 @@ import java.util.UUID;
 public class Customer {
 
     private final UUID customerId;
-    private final String customerName;
-    private final CustomerType customerType;
+    private String customerName;
+    private CustomerType customerType;
 
     public Customer(UUID customerId, String customerName, CustomerType customerType) {
         isValidCustomerName(customerName);
         this.customerId = customerId;
         this.customerName = customerName;
         this.customerType = customerType;
+    }
+
+    public void updateCustomerInfo(CustomerRequest customerRequest){
+        isValidCustomerName(customerRequest.getCustomerName());
+        this.customerName = customerRequest.getCustomerName();
+        this.customerType = customerRequest.getCustomerType();
     }
 
     private void isValidCustomerName(String customerName) {
