+++ src/main/java/com/zerozae/voucher/service/customer/CustomerService.java
@@ -6,6 +6,7 @@ import com.zerozae.voucher.dto.customer.CustomerRequest;
 import com.zerozae.voucher.dto.customer.CustomerResponse;
 import com.zerozae.voucher.exception.ErrorMessage;
 import com.zerozae.voucher.repository.customer.CustomerRepository;
+import com.zerozae.voucher.repository.customer.FileCustomerRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
